// src/types/shop/address.ts
export interface IShopAddress {
  id?: string;
  shopProfile?: string;
  phone?: string;
  fullName?: string;
  city?: string;
  address?: string;
  subAddress?: string;
  label?: string;
  isDefault?: string;
  status?: string;
  isDeleted?: boolean;
  deletedAt?: string;
  dateCreated?: string;
  dateUpdated?: string;
}